pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[project]
name = "mcp-poc"
version = "0.1.0"
description = "MCP PoC: chain vs agent vs MCP server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=0.3.0",
    "chromadb>=0.6.0",
    "mlflow>=2.20.0",
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
